CBLS

Documentation for CBLS.

CBLS.AllDifferent

CBLS.AllEqual

CBLS.AllEqualParam

CBLS.AlwaysTrue

CBLS.DiscreteSet

CBLS.DistDifferent

CBLS.Eq

CBLS.Error

CBLS.MOIAllDifferent

CBLS.MOIAllEqual

CBLS.MOIAllEqualParam

CBLS.MOIAlwaysTrue

CBLS.MOIDistDifferent

CBLS.MOIEq

CBLS.MOIError

CBLS.MOIOrdered

CBLS.MOIPredicate

CBLS.MOISumEqualParam

CBLS.Optimizer

CBLS.Optimizer

CBLS.Ordered

CBLS.Predicate

CBLS.ScalarFunction

CBLS.SumEqualParam

Base.copy

Base.copy

JuMP.build_variable

MathOptInterface.Utilities.supports_default_copy_to

MathOptInterface.add_constraint

MathOptInterface.add_constraint

MathOptInterface.add_variable

MathOptInterface.empty!

MathOptInterface.get

MathOptInterface.is_empty

MathOptInterface.optimize!

MathOptInterface.set

MathOptInterface.set

MathOptInterface.set

MathOptInterface.supports_constraint

CBLS.AllDifferent — Type

Global constraint ensuring that all the values of a given configuration are unique.

@constraint(model, X in AllDifferent())

CBLS.AllEqual — Type

Global constraint ensuring that all the values of X are all equal.

@constraint(model, X in AllEqual())

CBLS.AllEqualParam — Type

Global constraint ensuring that all the values of X are all equal to a given parameter param.

@constraint(model, X in AllEqualParam(param))

CBLS.AlwaysTrue — Type

Always return true. Mainly used for testing purpose.

@constraint(model, X in AlwaysTrue())

CBLS.DiscreteSet — Type

DiscreteSet(values)

CBLS.DistDifferent — Type

Local constraint ensuring that, given a vector X of size 4, |X[1] - X[2]| ≠ |X[3] - X[4]|).

@constraint(model, X in DistDifferent())

CBLS.Eq — Type

Equality between two variables.

@constraint(model, X in Eq())

CBLS.Error — Type

Error{F <: Function} <: JuMP.AbstractVectorSet

The solver will compute a straightforward error function based on the concept. To run the solver efficiently, it is possible to provide an error function err instead of concept. err must return a nonnegative real number.

@constraint(model, X in Error(err))

CBLS.MOIAllDifferent — Type

MOIAllDifferent <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIAllEqual — Type

MOIAllEqual <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIAllEqualParam — Type

MOIAllEqualParam{T <: Number} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

param::T: DESCRIPTION

dimension::Int: DESCRIPTION

MOIAllEqualParam(param, dim = 0) = begin #= none:5 =# new{typeof(param)}(param, dim) end: DESCRIPTION

CBLS.MOIAlwaysTrue — Type

MOIAlwaysTrue <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIDistDifferent — Type

MOIDistDifferent <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIEq — Type

MOIEq <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIError — Type

MOIError{F <: Function} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

f::F: DESCRIPTION

dimension::Int: DESCRIPTION

MOIError(f, dim = 0) = begin #= none:5 =# new{typeof(f)}(f, dim) end: DESCRIPTION

CBLS.MOIOrdered — Type

MOIOrdered <: MOI.AbstractVectorSet

DOCSTRING

CBLS.MOIPredicate — Type

MOIPredicate{F <: Function} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

f::F: DESCRIPTION

dimension::Int: DESCRIPTION

MOIPredicate(f, dim = 0) = begin #= none:5 =# new{typeof(f)}(f, dim) end: DESCRIPTION

CBLS.MOISumEqualParam — Type

MOISumEqualParam{T <: Number} <: MOI.AbstractVectorSet

DOCSTRING

Arguments:

param::T: DESCRIPTION

dimension::Int: DESCRIPTION

MOISumEqualParam(param, dim = 0) = begin #= none:5 =# new{typeof(param)}(param, dim) end: DESCRIPTION

CBLS.Optimizer — Type

Optimizer <: MOI.AbstractOptimizer

DOCSTRING

Arguments:

solver::Solver: DESCRIPTION

status::MOI.TerminationStatusCode: DESCRIPTION

options::Options: DESCRIPTION

CBLS.Optimizer — Type

Optimizer(model = Model(); options = Options())

DOCSTRING

CBLS.Ordered — Type

Global constraint ensuring that all the values of x are ordered.

@constraint(model, X in Ordered())

CBLS.Predicate — Type

Predicate{F <: Function} <: JuMP.AbstractVectorSet

Assuming X is a (collection of) variables, concept a boolean function over X, and that a model is defined. In JuMP syntax we can create a constraint based on concept as follows.

@constraint(model, X in Predicate(concept))

CBLS.ScalarFunction — Type

ScalarFunction{F <: Function, V <: Union{Nothing, VOV}} <: MOI.AbstractScalarFunction

A container to express any function with real value in JuMP syntax. Used with the @objective macro.

Arguments:

f::F: function to be applied to X

X::V: a subset of the variables of the model.

Given a model, and some (collection of) variables X to optimize. an objective function f can be added as follows. Note that only Min for minimization us currently defined. Max will come soon.

# Applies to all variables in order of insertion.
# Recommended only when the function argument order does not matter.
@objective(model, ScalarFunction(f))

# Generic use
@objective(model, ScalarFunction(f, X))

CBLS.SumEqualParam — Type

Global constraint ensuring that the sum of the values of X is equal to a given parameter param.

@constraint(model, X in SumEqualParam(param))

Base.copy — Method

Base.copy(set::MOIError) = begin

DOCSTRING

Base.copy — Method

Base.copy(set::DiscreteSet) = begin

DOCSTRING

JuMP.build_variable — Method

JuMP.build_variable(::Function, info::JuMP.VariableInfo, set::T) where T <: MOI.AbstractScalarSet

DOCSTRING

Arguments:

``: DESCRIPTION

info: DESCRIPTION

set: DESCRIPTION

MathOptInterface.Utilities.supports_default_copy_to — Method

Copy constructor for the optimizer

MathOptInterface.add_constraint — Method

MOI.add_constraint(optimizer::Optimizer, vars::MOI.VectorOfVariables, set::MOIError)

DOCSTRING

Arguments:

optimizer: DESCRIPTION

vars: DESCRIPTION

set: DESCRIPTION

MathOptInterface.add_constraint — Method

MOI.add_constraint(optimizer::Optimizer, v::SVF, set::DiscreteSet{T}) where T <: Number

DOCSTRING

Arguments:

optimizer: DESCRIPTION

v: DESCRIPTION

set: DESCRIPTION

MathOptInterface.add_variable — Method

MOI.add_variable(model::Optimizer) = begin

DOCSTRING

MathOptInterface.empty! — Method

MOI.empty!(opt) = begin

DOCSTRING

MathOptInterface.get — Method

MOI.get(::Optimizer, ::MOI.SolverName) = begin

DOCSTRING

MathOptInterface.is_empty — Method

MOI.is_empty(model::Optimizer) = begin

DOCSTRING

MathOptInterface.optimize! — Method

MOI.optimize!(model::Optimizer)

MathOptInterface.set — Function

MOI.set(::Optimizer, ::MOI.Silent, bool = true) = begin

DOCSTRING

Arguments:

``: DESCRIPTION

``: DESCRIPTION

bool: DESCRIPTION

MathOptInterface.set — Method

MOI.set(model::Optimizer, p::MOI.RawParameter, value)

Set a RawParameter to value

MathOptInterface.set — Method

MOI.set(model::Optimizer, ::MOI.TimeLimitSec, value::Union{Nothing,Float64})

Set the time limit

MathOptInterface.supports_constraint — Method

MOI.supports_constraint(::Optimizer, ::Type{VOV}, ::Type{MOIError}) = begin

DOCSTRING

Arguments:

``: DESCRIPTION

``: DESCRIPTION

``: DESCRIPTION